add respawn component
and correctGameManagerComponent.h
