+++ CSC8508/GameManagerComponent.h
@@ -1,10 +1,8 @@
-#pragma once
 #ifndef GAME_MANAGER_COMPONENT_H
 #define GAME_MANAGER_COMPONENT_H
 
 #include "IComponent.h"
 #include <iostream>
-#include <vector>
 
 namespace NCL::CSC8508 {
 	class GameManagerComponent : public IComponent {
@@ -12,7 +10,6 @@ namespace NCL::CSC8508 {
 		int quota;
 		int bankedCurrency;
 		int terminationFee;
-		std::vector<Vector3> spawnPoints;
 
 		static GameManagerComponent* instance;
 
@@ -50,14 +47,6 @@ namespace NCL::CSC8508 {
 
 		void OnMissionFailure() {
 			std::cout << "Mission failed! You lost!" << std::endl;
-		}
-
-		void SetSpawnPoints(const std::vector<Vector3>& points) {
-			spawnPoints = points;
-		}
-
-		const std::vector<Vector3>& GetSpawnPoints() const {
-			return spawnPoints;
 		}
 	};
 }
